OpenDRT v1.0.0b28 - Expose controls for the "Rendering space" (How "desaturated" we are from P3, with custom weights.)
I think I have a much better solution for the blues now. I've decided to expose the controls for the "rendering space" since it does allow some interesting creative options. I've exposed the controls on the purity tab and called the main control "color contrast" - though it could probably be called "saturation" since boosting it increases colorfulness (and introduces top end display cube escapes which are then brought back in control by the density parameters)...

And I've significantly reduced the "weight B", which was previously quite high - this had nice appearance in blue lights but was causing pretty significant out of gamut issues at the bottom end which is what you were picking up on in that frame with the blue light on the carnival ride (thank you for spotting this and questioning it) .. i think where it is not is much improved, and the complexity is not increased.

I also decided to add purity compress controls for R G and B - it's like a multiply on top of the "global" purity compress controls - and i think it would be another useful parameter to adjust. If you wanted to boost B density, this might not be necessary anymore as you can instead tweak the purity compression for that region and probably get something that is better.
Actually, reducing B weight also changes the luminance of yellow which is not great... so i brought back the CMY density weights so its possible to compensate, and reworked the presets to use both approaches.